Add files via upload
Easily compiled with auto-py-to-exe. Exe size with all required libraries is about 42mb. The app is designed to be installed per user using inno-script as it needs a folder to write,read and save the file_index json and the tags.json files. Otherwise, run the python script or the exe from within a folder. I use inno script to compile the exe into an installer which is the best method for ease of installation and proper functionality. This application has been tested in a small business environment and personal PC's without issues handling up to a total of 50,000 files. 5,000 files per directory X 10. This app is limited to user adding no more than 10 directories total. It handles them with lightning quick speed as well.